Vtable tp and kins try3

+++ src/emc/nml_intf/emc.hh
@@ -552,7 +552,7 @@ int emcSetMaxFeedOverride(double maxFeedScale);
 int emcSetupArcBlends(int arcBlendEnable,
         int arcBlendFallbackEnable,
         int arcBlendOptDepth,
-        double arcBlendGapCycles,
+        int arcBlendGapCycles,
         double arcBlendRampFreq);
 
 extern int emcUpdate(EMC_STAT * stat);
